+++ test/support/test_modules/hidden.ex
@@ -1,0 +1,8 @@
+defmodule Livebook.TestModules.Hidden do
+  @moduledoc false
+
+  def visible, do: :ok
+
+  @doc false
+  def hidden, do: :ok
+end